+++ libraries/TTBAS_LIB/tGraphicDev.h
@@ -1,5 +1,5 @@
 //
-// 豊四季Tiny BASIC for Arduino STM32 グラフィック描画デバイス
+// 豊四季Tiny BASIC for Arduino STM32 Graphic drawing device
 // 2017/07/19 by たま吉さん
 //
 
@@ -14,20 +14,20 @@ const uint8_t* tv_getFontAdr() ;
 
 class tGraphicDev {
   protected:
-    uint16_t gwidth;             // スクリーングラフィック横サイズ
-    uint16_t gheight;            // スクリーングラフィック縦サイズ
+    uint16_t gwidth;             // Screen graphic horizontal size
+    uint16_t gheight;            // Screen graphic vertical size
 
   public:
      void  init();
-    inline const uint8_t *getfontadr() { return tv_getFontAdr()+3; };   // フォントアドレスの参照
-    virtual uint16_t getGWidth();       // グラフックスクリーン横幅取得
-    virtual uint16_t getGHeight();      // グラフックスクリーン縦幅取得
+    inline const uint8_t *getfontadr() { return tv_getFontAdr()+3; };   // Font address pointer
+    virtual uint16_t getGWidth();       // Graphics screen width acquisition
+    virtual uint16_t getGHeight();      // Graphics screen vertical width acquisition
 
-    // グラフィック描画
+    // Graphic drawing
     virtual void pset(int16_t x, int16_t y, uint8_t c);
     virtual void line(int16_t x1, int16_t y1, int16_t x2, int16_t y2, uint8_t c);
-    virtual void circle(int16_t x, int16_t y, int16_t r, uint8_t c, int8_t f);
-    virtual void rect(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t c, int8_t f);
+    virtual void circle(int16_t x, int16_t y, int16_t r, uint8_t c, int f);
+    virtual void rect(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t c, int f);
     //virtual void cscroll(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t d);
     virtual void gscroll(int16_t x, int16_t y, int16_t w, int16_t h, uint8_t mode);
     virtual void set_gcursor(uint16_t, uint16_t);
@@ -36,4 +36,3 @@ class tGraphicDev {
 };
 
 #endif
-
